+++ src/main/java/uk/gov/ida/verifyserviceprovider/configuration/VerifyServiceProviderMetadataConfiguration.java
@@ -1,7 +1,0 @@
-package uk.gov.ida.verifyserviceprovider.configuration;
-
-import uk.gov.ida.saml.metadata.MetadataResolverConfiguration;
-
-public interface VerifyServiceProviderMetadataConfiguration extends MetadataResolverConfiguration {
-    boolean shouldLoadTrustStoreFromResources();
-}