+++ src/test/java/unit/uk/gov/ida/verifyserviceprovider/configuration/MsaMetadataConfigurationTest.java
@@ -1,16 +1,22 @@
 package unit.uk.gov.ida.verifyserviceprovider.configuration;
 
+import org.junit.Rule;
 import org.junit.Test;
+import org.junit.rules.ExpectedException;
 import uk.gov.ida.verifyserviceprovider.configuration.MsaMetadataConfiguration;
+import unit.uk.gov.ida.verifyserviceprovider.VerifyServiceProviderConfigurationTest;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static uk.gov.ida.verifyserviceprovider.configuration.ConfigurationConstants.MSA_JERSEY_CLIENT_NAME;
 import static uk.gov.ida.verifyserviceprovider.utils.DefaultObjectMapper.OBJECT_MAPPER;
 
 public class MsaMetadataConfigurationTest {
+    @Rule
+    public final ExpectedException expectedException = ExpectedException.none();
+
     @Test
     public void shouldSetDefaultConfigValuesWhenNotProvided() throws Exception {
-        String configurationAsString = "{\"uri\": \"http://some-msa-uri\"}";
+        String configurationAsString = "{\"uri\": \"http://some-msa-uri\", \"expectedEntityId\": \"foo\"}";
 
         MsaMetadataConfiguration actualConfiguration = OBJECT_MAPPER.readValue(configurationAsString, MsaMetadataConfiguration.class);
 
@@ -18,5 +24,13 @@ public class MsaMetadataConfigurationTest {
         assertThat(actualConfiguration.getMaxRefreshDelay()).isEqualTo(600000);
         assertThat(actualConfiguration.getJerseyClientConfiguration()).isNotNull();
         assertThat(actualConfiguration.getJerseyClientName()).isEqualTo(MSA_JERSEY_CLIENT_NAME);
+    }
+
+    @Test
+    public void shouldNotAllowEmptyExpectedEntityId() throws Exception {
+        expectedException.expectMessage("Missing required creator property 'expectedEntityId'");
+        String configurationAsString = "{\"uri\": \"http://some-msa-uri\"}";
+
+        OBJECT_MAPPER.readValue(configurationAsString, MsaMetadataConfiguration.class);
     }
 }